modif back front

+++ backend/controllers/userctrl.js
@@ -17,7 +17,6 @@ exports.getUser = (req, res, next) => {
   User.findOne({ userId: req.params.id })
     .then((user) => {
       res.status(200).json(user)
-      console.log(user);
     })
     .catch((error) => res.status(400).json({ error }));
 }
@@ -25,6 +24,7 @@ exports.getUser = (req, res, next) => {
 allows you to create a user in the DB
 */
 exports.createUser = (req, res, next) => {
+  // console.log(req.body);
   const userObject = req.body;
   delete userObject._id;
   let user = '';
@@ -33,25 +33,29 @@ exports.createUser = (req, res, next) => {
       ...userObject,
       imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`,
     });
+    console.log(user);
   } else {
     user = new User({
       ...userObject,
     });
+    console.log(user);
   }
   user.save()
     .then(() => {
       res.status(201).json({ message: "Votre profil a bien été créé" });
       console.log(user);
     })
-    .catch(() => res.status(400).json({ message: "Ce pseudo est déjà utiliseé" }));
+    .catch(() => res.status(400).json({ message: "un erreur c'est produite" }));
 };
 /*
 allows you to update a user in the DB by the user Id
 */
 exports.updateUser = (req, res, next) => {
+  console.log(req.body);
+  console.log(req.file);
   const userObject = req.file ?
     {
-      ...JSON.parse(req.body.user),
+      ...req.body.user,
       imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename
         }`,
     }
@@ -64,16 +68,29 @@ exports.updateUser = (req, res, next) => {
 allows you to delete a user in the DB by thr user ID
 */
 exports.deleteUser = (req, res, next) => {
+  console.log(req.params);
   User.findOne({ userId: req.params.id })
     .then((user) => {
-      const filename = user.imageUrl.split("/images")[1];
-      fs.unlink(`images/${filename}`, () => {
+      if (user.imageUrl !== null) {
+        console.log(user);
+        const filename = user.imageUrl.split("/images")[1];
+        console.log(filename);
+        fs.unlink(`images/${filename}`, () => {
+          User.deleteOne({ userId: req.params.id })
+            .then(() =>
+              res.status(200).json({ message: "Votre profil a été supprimée" })
+            )
+            .catch((error) => res.status(400).json({ error }));
+        });
+      }
+      else {
         User.deleteOne({ userId: req.params.id })
-          .then(() =>
-            res.status(200).json({ message: "Votre profil a été supprimée" })
-          )
+          .then(() => {
+            res.status(200).json({ message: "Le compte utilisateur supprimé" })
+          })
           .catch((error) => res.status(400).json({ error }));
-      });
+      }
+
     })
     .catch((error) => res.status(500).json({ error }));
 };
